Powershell and fix (#23)
* Powershell and fix

Pretty much every single language had a bug in it.
When entering 1 as the input, it would output 0 as the peak number (because the peak number variable was always initialized with 0). However, this is not correct as the peak number reached would be 1.
I fixed that.
I also updated some of my implementations (C++ was absolute garbage).
I decided to not add credit in every single file.
Oh and I also did powershell.

* Revert fix for implementations I didn't write, and bug fix for my implementations

Now all my implementations (except for golfed C and golfed JS) validate user input.

+++ c/c.c
@@ -1,5 +1,5 @@
 // Written with <3 by @DrParanoya on 08/23/2025
-// Golfed Collatz Conjecture in Ansi-C (165 characters, 179 with -includetime.h)
+// Golfed Collatz Conjecture in Ansi-C (154 characters, 168 with -includetime.h)
 // gcc c.c -ansi -includetime.h
 
-main(){uint64_t n,s=scanf("%llu",&n),p=s;auto a=clock();for(;n!=1;){n=n%2?n*3+1:n/2;++s;p=n>p?n:p;}printf("%llu %llu %f",--s,p,((float)(clock()-a))/CLOCKS_PER_SEC);}
+main(){int n,s=scanf("%d",&n),p=s;auto a=clock();for(;n!=1;){n=n%2?n*3+1:n/2;++s;p=n>p?n:p;}printf("%d %d %f",--s,p,((float)(clock()-a))/CLOCKS_PER_SEC);}
